% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{Fizz Buzz Child's Game}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{A positive integer}
}
\value{
n or "Fizz", "Buzz", "Fizz Buzz" based on divisibility
}
\description{
This function will output a vector of the first n terms in the child's game Fizz Buzz.
The input of the function is a positive integer n and the output will be a vector of
characters. The goal is to count as high as you can, but for any number evenly divisible
by three, substitute "Fizz", and any number divisible by five substitute "Buzz", and if
it is divisible by both, substitute "Fizz Buzz". So, the sequence will look like 1, 2,
Fizz, 4, Buzz, Fizz, 7, 8, Fizz...
}
\examples{
FizzBuzz(50)
FizzBuzz(15)
}
