% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Flagstaff weather phenomena at Pulliam airport from 1950-2019.}
\format{
a data frame with 25449 observations of 5 variables
\describe{
  \item{DATE}{The date of observation as a POSIX date format.}
  \item{PRCP}{Precipitation in inches.}
  \item{SNOW}{Snowfall in inches.}
  \item{TMAX}{Maximum temperature in degrees Farenheit.}
  \item{MinTemp}{Minimum temperature in degrees Farenheit.}
}
}
\source{
\url{https://raw.githubusercontent.com/dereksonderegger/444/master/data-raw/Pulliam_Airport_Weather_Station.csv}
}
\usage{
Flagstaff_Weather
}
\description{
Flagstaff weather phenomena at Pulliam airport from 1950-2019.
}
\keyword{datasets}
